Moved milestones fetch to main content
-moved milestones fetch to main content
-passed milestones data to milestones and codewars components in props
fixed doughnut chart, added percentage
