+++ client/src/pages/Trainee_Dashboard/componets/Milestones.js
@@ -10,110 +10,83 @@ import Paper from "@mui/material/Paper";
 import CircularProgress from "@mui/material/CircularProgress";
 import { Typography } from "@mui/material";
 
-export default function Milestones() {
-  const [milestonesData, setMilestonesData] = useState([]);
-  const [isLoading, setIsLoading] = useState(true);
-  const [currentMilestoneData, setCurrentMilestoneData] = useState(null);
-
-  useEffect(() => {
-    const fetchData = async () => {
-      try {
-        setIsLoading(true);
-        const instant = axios.configAxios();
-        const response = await instant.get("/milestones");
-        setMilestonesData(response.data);
-        setIsLoading(false);
-      } catch (error) {
-        console.error("Error fetching milestones:", error);
-        setIsLoading(false);
-      }
-    };
-
-    fetchData();
-  }, []);
-
-  useEffect(() => {
-    setIsLoading(true);
-    const instant = axios.configAxios();
-    instant
-      .get("/current-milestone")
-      .then((data) => {
-        setIsLoading(false);
-        setCurrentMilestoneData(data);
-      })
-      .catch((error) => {
-        setIsLoading(false);
-      });
-  }, []);
+export default function Milestones({ milestonesData, isLoading }) {
+  const [currentMilestoneData] = useState(null);
 
   return (
     <div>
-      <Typography variant="h4" gutterBottom style={{textAlign: "center" , color: "#D82929", fontWeight: "bold"}}>
-      Milestone Table 
+      <Typography
+        variant="h4"
+        gutterBottom
+        style={{ textAlign: "center", color: "#D82929", fontWeight: "bold" }}
+      >
+        Milestone Table
       </Typography>
-      <Typography variant="body1" paragraph style={{textAlign: "center"}}>
-      Check your milestones and performance expectations with dates.
+      <Typography variant="body1" paragraph style={{ textAlign: "center" }}>
+        Check your milestones and performance expectations with dates.
       </Typography>
-    <TableContainer component={Paper}>
-      {isLoading ? (
-        <CircularProgress />
-      ) : (
-        <Table sx={{ minWidth: 650 }} aria-label="simple table">
-          <TableHead>
-            <TableRow sx={{ backgroundColor: "#d5d4d4" }}>
-              <TableCell>No</TableCell>
-              <TableCell>Milestone</TableCell>
-              <TableCell>Start Date</TableCell>
-              <TableCell>Codewars</TableCell>
-              <TableCell>Codility</TableCell>
-              <TableCell>Pull Reqs</TableCell>
-              <TableCell>Attendance</TableCell>
-            </TableRow>
-          </TableHead>
-          <TableBody sx={{ backgroundColor: "#fafafa" }}>
-            {milestonesData.map((row, index) => (
-              <TableRow
-                key={row.id}
-                sx={{
-                  "&:last-child td, &:last-child th": { border: 0 },
-                  "&:hover": {
-                    backgroundColor: index % 2 === 0 ? "#f0f0f0" : "#e0e0e0",
-                  },
-                  ...(currentMilestoneData &&
-                    currentMilestoneData.data.name === row.milestone && {
-                      backgroundColor: "#ffcccc",
-                    }),
-                }}
-              >
-                <TableCell component="th" scope="row">
-                  {row.id}
-                </TableCell>
-                <TableCell>{row.milestone}</TableCell>
-                <TableCell>
-                  {new Date(row.startDate).toLocaleDateString()}
-                </TableCell>
-                <TableCell>
-                  {row.factors.find((factor) => factor.name === "Codewars")
-                    ?.value || 0}
-                </TableCell>
-                <TableCell>
-                  {row.factors.find((factor) => factor.name === "Codility")
-                    ?.value || 0}
-                </TableCell>
-                <TableCell>
-                  {row.factors.find((factor) => factor.name === "Pulls")
-                    ?.value || 0}
-                </TableCell>
-                <TableCell>{`${
-                  row.factors.find((factor) => factor.name === "Attendance")
-                    ?.value || 0
-                }%`}</TableCell>
+      <TableContainer component={Paper}>
+        {isLoading ? (
+          <CircularProgress />
+        ) : (
+          <Table sx={{ minWidth: 650 }} aria-label="simple table">
+            <TableHead>
+              <TableRow sx={{ backgroundColor: "#d5d4d4" }}>
+                <TableCell>No</TableCell>
+                <TableCell>Milestone</TableCell>
+                <TableCell>Start Date</TableCell>
+                <TableCell>Codewars</TableCell>
+                <TableCell>Codility</TableCell>
+                <TableCell>Pull Reqs</TableCell>
+                <TableCell>Attendance</TableCell>
               </TableRow>
-            ))}
-          </TableBody>
-        </Table>
-      )}
-    </TableContainer>
+            </TableHead>
+            <TableBody sx={{ backgroundColor: "#fafafa" }}>
+              {milestonesData !== null &&
+                milestonesData.map((row, index) => (
+                  <TableRow
+                    key={row.id}
+                    sx={{
+                      "&:last-child td, &:last-child th": { border: 0 },
+                      "&:hover": {
+                        backgroundColor:
+                          index % 2 === 0 ? "#f0f0f0" : "#e0e0e0",
+                      },
+                      ...(currentMilestoneData &&
+                        currentMilestoneData.data.name === row.milestone && {
+                          backgroundColor: "#ffcccc",
+                        }),
+                    }}
+                  >
+                    <TableCell component="th" scope="row">
+                      {row.id}
+                    </TableCell>
+                    <TableCell>{row.milestone}</TableCell>
+                    <TableCell>
+                      {new Date(row.startDate).toLocaleDateString()}
+                    </TableCell>
+                    <TableCell>
+                      {row.factors.find((factor) => factor.name === "Codewars")
+                        ?.value || 0}
+                    </TableCell>
+                    <TableCell>
+                      {row.factors.find((factor) => factor.name === "Codility")
+                        ?.value || 0}
+                    </TableCell>
+                    <TableCell>
+                      {row.factors.find((factor) => factor.name === "Pulls")
+                        ?.value || 0}
+                    </TableCell>
+                    <TableCell>{`${
+                      row.factors.find((factor) => factor.name === "Attendance")
+                        ?.value || 0
+                    }%`}</TableCell>
+                  </TableRow>
+                ))}
+            </TableBody>
+          </Table>
+        )}
+      </TableContainer>
     </div>
   );
 }
